lib: use default repl writer
The REPLServer.writer signature is synchronous which means we cannot
aquire any more information about the remote result, so for things like
objects we have to bail and just report that its an object without being
able to find out any more information about it.

This uses the default writer, and moves the value evaluation back into
REPLServer.eval, which is asynchronous and will let us dig as deep as
we need.