lib: filter unavailable debugger websocket urls
Debugging targets from /json/list are always given a webSocketDebuggingUrl even
when it should not available, which can result in fatal errors when trying to
click a non-connectable tab.

This fixes that by only transforming that property if the target already has a
`webSocketDebuggerUrl` propery or if the target is an upstream connection.