lib: ensure debug port is open before spawn
If the port given to browser is occupied we would get stuck in an
loop waiting for the target tab to be available.

This addresses that by testing if the port is open before attempting to
spawn a browser. If the port is not open an error will be yielded.