lib: normalize compiler output paths
When using at least one folder on windows, for example src/index.js. The
script key (pathname) would be assigned as `src/index.js` for the
scripts map. While in the server it would be normalized to
`src\index.js` failing to find the script mapping.

This makes it so both are normalized therefore the pathname can be
correctly found inside the scripts and served by the http server.