bin: use a single env variable pair for browsers
Currently there's a plethora of redundant browser environment variables.

This reduces them down to a single `BROWSER` and `BROWSER_FLAGS` pair,
the first represents the default executable path if no browser name or
executable path is given with the option, the latter additional command
line arguments.

As a side effect, the argument for `--browser` is now optional.