lib: remove listener duplication in `hotpatch`
There's a possibility for  events to be added more than once, this
assumes that client is fresh every time `connect` is emitted. Which is
not true. This results in multiple handlers being added in the case of a
re-connect.

This fixes that by flattening the listeners, adding them only once per
object.