lib: rename plugin functions into verb forms
Some of the plugins are not named in verb forms, while others are. In
terms of semantics there are valid reasons to rename everything into
verbs, like `browser`, should open a new tab if the browser already
exists.

This renames all plugin functions into verb forms of their current name.