lib: let browser normalize urls for us
Currently we're attempting to normalize urls before passing them to the
browser, however this is somewhat redundant as we can instead do it the
other way around and get the normalized url from the browser.

This passes the url as-is to the browser, then sets the url as the url
of the first matching target it finds once a connection can be made to
the browser.