lib: reuse profiles based on command and port
New browser profiles are created every time `browse` is used, this is
not strictly needed as we do a port check before spawning the browser.
So creating a profile on a per port basis is enough.

This replaces the random temporary profile paths with predictable
profile paths based on the browser port and command. This has two major
benefits, the first is that it reduces start-up time by quite a bit. The
second being that window positions are now remembered, which is really
great when you have the editor on one side of the monitor and want the
browser to always take the other half.