test: refactor and use environment variables
We iterate over all the possible permutations of the tests, this leads
to very deeply nested and somewhat unreadable tests. In addition these
tests are very slow with the entire test suite taking an average of 20
minutes.

This refactors the tests to use block scoping, and uses environment
variables instead of enumeration.

Using environment variables will allow us to run tests in parallel with
`make -j`, and also selectivly only run failing tests.