doc: use banner image from media repository
We are hard-linking to a png version of the old logo with GitHub's
files, which is tedious to update as we iterate on the branding.

This changes the image url to point to the svg from the repository
through rawgit in order to get a correct mime-type. This will make
iterating on said image easier across the different repositories.